+++ templates/data_connector/src/api/index.ts
@@ -1,4 +1,0 @@
-export * from "./connect_client";
-export * from "./fetch_data_table";
-export * from "./data_source";
-export * from "./oauth";