+++ templates/data_connector/src/intents/data_connector/app.tsx
@@ -3,9 +3,9 @@ import { AppUiProvider } from "@canva/app-ui-kit";
 import type { RenderSelectionUiRequest } from "@canva/intents/data";
 import { ErrorBoundary } from "react-error-boundary";
 import { createHashRouter, RouterProvider } from "react-router-dom";
-import { ContextProvider } from "../../context";
-import { ErrorPage } from "../../pages";
-import { routes } from "../../routes";
+import { ContextProvider } from "../../context/app_context";
+import { ErrorPage } from "../../pages/error";
+import { routes } from "../../routes/routes";
 
 export const App = ({ request }: { request: RenderSelectionUiRequest }) => (
   <AppI18nProvider>