+++ templates/data_connector/src/pages/index.ts
@@ -1,4 +1,0 @@
-export * from "./error";
-export * from "./login";
-export * from "./select_source";
-export * from "./data_source_config";