+++ templates/gen_ai/src/intents/design_editor/app.tsx
@@ -2,9 +2,9 @@ import { AppI18nProvider } from "@canva/app-i18n-kit";
 import { AppUiProvider } from "@canva/app-ui-kit";
 import { ErrorBoundary } from "react-error-boundary";
 import { createHashRouter, RouterProvider } from "react-router-dom";
-import { ContextProvider } from "../../context";
-import { ErrorPage } from "../../pages";
-import { routes } from "../../routes";
+import { ContextProvider } from "../../context/app_context";
+import { ErrorPage } from "../../pages/error";
+import { routes } from "../../routes/routes";
 
 export const App = () => (
   <AppI18nProvider>