+++ templates/gen_ai/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./obscenity_filter";